requirements.txt
phidata
yfinance
duckduckgo-search
ollama
fastapi
openmeteo-requests